refactor(components): pagination
move the `pagination` from `components` to 'core/ui'
